add functionality of deleting item from meal plan, add much better way of filtering items, check if recipe already exist in the meal plan

+++ src/store/slices/mealPlanSlice.ts
@@ -3,6 +3,7 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { mealPLan } from 'utils/constants/mealPlan.constants'
 
 import { WeekPlanType, DishType } from 'types/MealPlan'
+import { isWeakSet } from 'util/types';
 
 type MealPlanState = {
    weekPlan: WeekPlanType[];
@@ -23,30 +24,49 @@ export const mealPlanSlice = createSlice({
    reducers: {
       addRecipeIntoMeal: {
          reducer: (state, { payload }: PayloadAction<PayloadAddRecipeType>): void => {
-            state.weekPlan = state.weekPlan.map(dayPlan => {
-               if (dayPlan.id === payload.idWeek) {
-                  dayPlan.dishes.push(payload.plannedRecipe)
-               }
+            const currentWeek = state.weekPlan.find(week => week.idWeek === payload.idWeek)
 
-               return dayPlan
-            })
+            if (currentWeek) {
+               const isRecipeAlreadyExist = currentWeek.dishes.some(item => item.idDish === payload.plannedRecipe.idDish)
 
-            localStorage.setItem('weekPlan', JSON.stringify(state.weekPlan))
+               if (isRecipeAlreadyExist) return
+
+               state.weekPlan = state.weekPlan.map(dayPlan => {
+                  if (dayPlan.idWeek === payload.idWeek) {
+                     dayPlan.dishes.push(payload.plannedRecipe)
+                  }
+   
+                  return dayPlan
+               })
+   
+               localStorage.setItem('weekPlan', JSON.stringify(state.weekPlan))
+            }
          },
          prepare: (idWeek: string, id: number, title: string, image: string) => {
             return {
                payload: {
                   idWeek,
-                  plannedRecipe: { id: String(id), title, image}
+                  plannedRecipe: { idDish: String(id), title, image }
                }
             }
          }
       },
-      deleteRecipeFromMealPlan: (state, action): void => {
-         console.log('test')  
+      deleteRecipeFromMealPlan: (state, { payload }: PayloadAction<{idDish: string, idWeek: string}>): void => {
+         state.weekPlan = state.weekPlan.map(week => {
+            if (week.idWeek === payload.idWeek) {
+               return {
+                  ...week,
+                  dishes: week.dishes.filter(dish => dish.idDish !== payload.idDish) 
+               }
+            } 
+
+            return week
+         })
+
+         localStorage.setItem('weekPlan', JSON.stringify(state.weekPlan))
       }
    }
 })
 
 export default mealPlanSlice.reducer
-export const { addRecipeIntoMeal } = mealPlanSlice.actions
+export const { addRecipeIntoMeal, deleteRecipeFromMealPlan } = mealPlanSlice.actions
